Store schemas as SQL (#403)
This changes the way table and index schemas are stored. Previously, we used to store the parsed representation (*database.TableInfo, *database.IndexInfo) as custom documents.
Instead, we now store the table information in SQL form and parse it at startup.
The motivation behind this change is to reduce the areas that could lead to a breaking change in the future. That way, if we decide to improve the internal representation of table and index information, the SQL representation won't change, neither the storage structure.
This also adds support for prepared statements.