Rework index encoding
Indexes are now encoded the same way whether there is a unique constraint
or not. The primary key is now always appended to the indexed values
and stored as the key of the index tree. The value of the index tree
now contains the offset of the primary key, stored as an unsigned varint.